Analise-REGINA-main/Super BI 2025/requirements.txt
pdfplumber>=0.11.0
pdfminer.six>=20240706
fastapi>=0.112.0
uvicorn>=0.30.0
pandas>=2.2.2
openpyxl>=3.1.4
pytesseract>=0.3.10
pytest>=8.2.0
pillow>=8.0.0
